Externalize InsightRestTemplateConfiguration and renaming some files.
