--- backend/src/main/java/ch/zuehlke/fullstack/ConnectZuehlke/apis/insight/service/RemoteInsightEmployeeService.java
@@ -1,104 +1,0 @@
-package ch.zuehlke.fullstack.ConnectZuehlke.apis.insight.service;
-
-import ch.zuehlke.fullstack.ConnectZuehlke.apis.insight.config.InsightProperties;
-import ch.zuehlke.fullstack.ConnectZuehlke.apis.insight.dto.EmployeeDto;
-import ch.zuehlke.fullstack.ConnectZuehlke.domain.Employee;
-import org.apache.http.auth.AuthSchemeProvider;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.NTCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.client.config.AuthSchemes;
-import org.apache.http.config.Registry;
-import org.apache.http.config.RegistryBuilder;
-import org.apache.http.impl.auth.BasicSchemeFactory;
-import org.apache.http.impl.auth.NTLMSchemeFactory;
-import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.*;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.http.converter.ByteArrayHttpMessageConverter;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.Arrays;
-import java.util.List;
-
-import static java.util.stream.Collectors.toList;
-import static org.springframework.http.HttpMethod.GET;
-
-@Service
-@Profile({"prod", "staging"})
-public class RemoteInsightEmployeeService implements InsightEmployeeService {
-    private final RestTemplate restTemplate;
-    private Logger logger = LoggerFactory.getLogger(RemoteInsightEmployeeService.class);
-
-    public RemoteInsightEmployeeService(RestTemplateBuilder restTemplateBuilder, InsightProperties insightProperties) {
-        this.restTemplate = restTemplateBuilder
-                .rootUri(insightProperties.getUrl())
-                .build();
-
-        String user = insightProperties.getAuthentication().getUsername();
-        String password = insightProperties.getAuthentication().getPassword();
-
-        logger.info("Authenticating on API {} with {}:", insightProperties.getUrl(), user);
-
-        CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-        credentialsProvider.setCredentials(AuthScope.ANY, new NTCredentials(user, password, null, null));
-
-        Registry<AuthSchemeProvider> authSchemeRegistry = RegistryBuilder.<AuthSchemeProvider>create()
-                .register(AuthSchemes.NTLM, new NTLMSchemeFactory())
-                .register(AuthSchemes.BASIC, new BasicSchemeFactory())
-                .build();
-
-        CloseableHttpClient httpclient = HttpClientBuilder.create()
-                .setDefaultCredentialsProvider(credentialsProvider)
-                .setDefaultAuthSchemeRegistry(authSchemeRegistry)
-                .build();
-
-        this.restTemplate.setRequestFactory(new HttpComponentsClientHttpRequestFactory(httpclient));
-
-        this.restTemplate.getMessageConverters().add(
-                new ByteArrayHttpMessageConverter());
-
-    }
-
-    @Override
-    public List<Employee> getEmployees() {
-        ResponseEntity<List<EmployeeDto>> response = this.restTemplate
-                .exchange("/employees", GET, null, new ParameterizedTypeReference<List<EmployeeDto>>() {
-                });
-
-        return response.getBody().stream()
-                .map(EmployeeDto::toEmployee)
-                .collect(toList());
-    }
-
-    @Override
-    public byte[] getEmployeePicture(int id) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setAccept(Arrays.asList(MediaType.APPLICATION_OCTET_STREAM));
-
-        HttpEntity<String> entity = new HttpEntity<>(headers);
-
-        ResponseEntity<byte[]> response = this.restTemplate
-                .exchange("/employees/" + id + "/picture?large=false", GET, entity, byte[].class, "1");
-        if (response.getStatusCode() == HttpStatus.OK) {
-            return response.getBody();
-        }
-        throw new IllegalStateException("Status code was not 200");
-    }
-
-    @Override
-    public Employee getEmployee(String code) {
-        ResponseEntity<EmployeeDto> response = this.restTemplate
-                .getForEntity("/employees/" + code, EmployeeDto.class);
-
-        return response.getBody().toEmployee();
-    }
-}
